# pkg/api/agent/private/display/v1/service.pb.goclay.go
// Code generated by protoc-gen-goclay. DO NOT EDIT.
// source: agent/private/display/v1/service.proto

/*
Package displayv1 is a self-registering gRPC and JSON+Swagger service definition.

It conforms to the github.com/utrack/clay/v3/transport Service interface.
*/
package displayv1

import (
	"bytes"
	"context"
	"encoding/base64"
	"fmt"
	"io/ioutil"
	"net/http"
	"net/url"
	"strings"

	"github.com/go-chi/chi"
	"github.com/go-openapi/spec"
	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
	"github.com/grpc-ecosystem/grpc-gateway/v2/utilities"
	"github.com/pkg/errors"
	"github.com/utrack/clay/v3/transport"
	"github.com/utrack/clay/v3/transport/httpclient"
	"github.com/utrack/clay/v3/transport/httpruntime"
	"github.com/utrack/clay/v3/transport/httpruntime/httpmw"
	"github.com/utrack/clay/v3/transport/httptransport"
	"github.com/utrack/clay/v3/transport/swagger"
	"google.golang.org/grpc"
)

// Update your shared lib or downgrade generator to v1 if there's an error
var _ = transport.IsVersion2

var _ = ioutil.Discard
var _ chi.Router
var _ runtime.Marshaler
var _ bytes.Buffer
var _ context.Context
var _ fmt.Formatter
var _ strings.Reader
var _ errors.Frame
var _ httpruntime.Marshaler
var _ http.Handler
var _ url.Values
var _ base64.Encoding
var _ httptransport.MarshalerError
var _ utilities.DoubleArray

// PrivateDisplayServiceDesc is a descriptor/registrator for the PrivateDisplayServiceServer.
type PrivateDisplayServiceDesc struct {
	svc  PrivateDisplayServiceServer
	opts httptransport.DescOptions
}

// NewPrivateDisplayServiceServiceDesc creates new registrator for the PrivateDisplayServiceServer.
// It implements httptransport.ConfigurableServiceDesc as well.
func NewPrivateDisplayServiceServiceDesc(svc PrivateDisplayServiceServer) *PrivateDisplayServiceDesc {
	return &PrivateDisplayServiceDesc{
		svc: svc,
	}
}

// RegisterGRPC implements service registrator interface.
func (d *PrivateDisplayServiceDesc) RegisterGRPC(s *grpc.Server) {
	RegisterPrivateDisplayServiceServer(s, d.svc)
}

// Apply applies passed options.
func (d *PrivateDisplayServiceDesc) Apply(oo ...transport.DescOption) {
	for _, o := range oo {
		o.Apply(&d.opts)
	}
}

// SwaggerDef returns this file's Swagger definition.
func (d *PrivateDisplayServiceDesc) SwaggerDef(options ...swagger.Option) (result []byte) {
	if len(options) > 0 || len(d.opts.SwaggerDefaultOpts) > 0 {
		var err error
		var s = &spec.Swagger{}
		if err = s.UnmarshalJSON(_swaggerDef_agent_private_display_v1_service_proto); err != nil {
			panic("Bad swagger definition: " + err.Error())
		}

		for _, o := range d.opts.SwaggerDefaultOpts {
			o(s)
		}
		for _, o := range options {
			o(s)
		}
		if result, err = s.MarshalJSON(); err != nil {
			panic("Failed marshal spec.Swagger definition: " + err.Error())
		}
	} else {
		result = _swaggerDef_agent_private_display_v1_service_proto
	}
	return result
}

// RegisterHTTP registers this service's HTTP handlers/bindings.
func (d *PrivateDisplayServiceDesc) RegisterHTTP(mux transport.Router) {
	chiMux, isChi := mux.(chi.Router)

	{
		// Handler for List, binding: GET /v1/displays
		var h http.HandlerFunc
		h = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			defer r.Body.Close()

			unmFunc := unmarshaler_goclay_PrivateDisplayService_List_0(r)
			rsp, err := _PrivateDisplayService_List_Handler(d.svc, r.Context(), unmFunc, d.opts.UnaryInterceptor)

			if err != nil {
				if err, ok := err.(httptransport.MarshalerError); ok {
					httpruntime.SetError(r.Context(), r, w, errors.Wrap(err.Err, "couldn't parse request"))
					return
				}
				httpruntime.SetError(r.Context(), r, w, err)
				return
			}

			if ctxErr := r.Context().Err(); ctxErr != nil && ctxErr == context.Canceled {
				w.WriteHeader(499) // Client Closed Request
				return
			}

			_, outbound := httpruntime.MarshalerForRequest(r)
			w.Header().Set("Content-Type", outbound.ContentType())
			err = outbound.Marshal(w, rsp)
			if err != nil {
				httpruntime.SetError(r.Context(), r, w, errors.Wrap(err, "couldn't write response"))
				return
			}
		})

		h = httpmw.DefaultChain(h)

		if isChi {
			chiMux.Method("GET", pattern_goclay_PrivateDisplayService_List_0, h)
		} else {
			mux.Handle(pattern_goclay_PrivateDisplayService_List_0, http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
				if r.Method != "GET" {
					w.WriteHeader(http.StatusMethodNotAllowed)
					return
				}
				h(w, r)
			}))
		}
	}

	{
		// Handler for SetEnabled, binding: POST /v1/displays/set-enable
		var h http.HandlerFunc
		h = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			defer r.Body.Close()

			unmFunc := unmarshaler_goclay_PrivateDisplayService_SetEnabled_0(r)
			rsp, err := _PrivateDisplayService_SetEnabled_Handler(d.svc, r.Context(), unmFunc, d.opts.UnaryInterceptor)

			if err != nil {
				if err, ok := err.(httptransport.MarshalerError); ok {
					httpruntime.SetError(r.Context(), r, w, errors.Wrap(err.Err, "couldn't parse request"))
					return
				}
				httpruntime.SetError(r.Context(), r, w, err)
				return
			}

			if ctxErr := r.Context().Err(); ctxErr != nil && ctxErr == context.Canceled {
				w.WriteHeader(499) // Client Closed Request
				return
			}

			_, outbound := httpruntime.MarshalerForRequest(r)
			w.Header().Set("Content-Type", outbound.ContentType())
			err = outbound.Marshal(w, rsp)
			if err != nil {
				httpruntime.SetError(r.Context(), r, w, errors.Wrap(err, "couldn't write response"))
				return
			}
		})

		h = httpmw.DefaultChain(h)

		if isChi {
			chiMux.Method("POST", pattern_goclay_PrivateDisplayService_SetEnabled_0, h)
		} else {
			mux.Handle(pattern_goclay_PrivateDisplayService_SetEnabled_0, http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
				if r.Method != "POST" {
					w.WriteHeader(http.StatusMethodNotAllowed)
					return
				}
				h(w, r)
			}))
		}
	}

}

type PrivateDisplayService_httpClient struct {
	c    *http.Client
	host string
}

// NewPrivateDisplayServiceHTTPClient creates new HTTP client for PrivateDisplayServiceServer.
// Pass addr in format "http://host[:port]".
func NewPrivateDisplayServiceHTTPClient(c *http.Client, addr string) *PrivateDisplayService_httpClient {
	if strings.HasSuffix(addr, "/") {
		addr = addr[:len(addr)-1]
	}
	return &PrivateDisplayService_httpClient{c: c, host: addr}
}

func (c *PrivateDisplayService_httpClient) List(ctx context.Context, in *ListRequest, opts ...grpc.CallOption) (*ListResponse, error) {
	mw, err := httpclient.NewMiddlewareGRPC(opts)
	if err != nil {
		return nil, err
	}

	path := pattern_goclay_PrivateDisplayService_List_0_builder(in)

	buf := bytes.NewBuffer(nil)

	m := httpruntime.DefaultMarshaler(nil)

	req, err := http.NewRequest("GET", c.host+path, buf)
	if err != nil {
		return nil, errors.Wrap(err, "can't initiate HTTP request")
	}
	req = req.WithContext(ctx)

	req.Header.Add("Accept", m.ContentType())

	req, err = mw.ProcessRequest(req)
	if err != nil {
		return nil, err
	}
	rsp, err := c.c.Do(req)
	if err != nil {
		return nil, errors.Wrap(err, "error from client")
	}
	defer rsp.Body.Close()

	rsp, err = mw.ProcessResponse(rsp)
	if err != nil {
		return nil, err
	}

	if rsp.StatusCode >= 400 {
		b, _ := ioutil.ReadAll(rsp.Body)
		return nil, errors.Errorf("%v %v: server returned HTTP %v: '%v'", req.Method, req.URL.String(), rsp.StatusCode, string(b))
	}

	ret := ListResponse{}

	err = m.Unmarshal(rsp.Body, &ret)

	return &ret, errors.Wrap(err, "can't unmarshal response")
}

func (c *PrivateDisplayService_httpClient) SetEnabled(ctx context.Context, in *SetEnabledRequest, opts ...grpc.CallOption) (*SetEnabledResponse, error) {
	mw, err := httpclient.NewMiddlewareGRPC(opts)
	if err != nil {
		return nil, err
	}

	path := pattern_goclay_PrivateDisplayService_SetEnabled_0_builder(in)

	buf := bytes.NewBuffer(nil)

	m := httpruntime.DefaultMarshaler(nil)

	if err = m.Marshal(buf, in); err != nil {
		return nil, errors.Wrap(err, "can't marshal request")
	}

	req, err := http.NewRequest("POST", c.host+path, buf)
	if err != nil {
		return nil, errors.Wrap(err, "can't initiate HTTP request")
	}
	req = req.WithContext(ctx)

	req.Header.Add("Accept", m.ContentType())

	req, err = mw.ProcessRequest(req)
	if err != nil {
		return nil, err
	}
	rsp, err := c.c.Do(req)
	if err != nil {
		return nil, errors.Wrap(err, "error from client")
	}
	defer rsp.Body.Close()

	rsp, err = mw.ProcessResponse(rsp)
	if err != nil {
		return nil, err
	}

	if rsp.StatusCode >= 400 {
		b, _ := ioutil.ReadAll(rsp.Body)
		return nil, errors.Errorf("%v %v: server returned HTTP %v: '%v'", req.Method, req.URL.String(), rsp.StatusCode, string(b))
	}

	ret := SetEnabledResponse{}

	err = m.Unmarshal(rsp.Body, &ret)

	return &ret, errors.Wrap(err, "can't unmarshal response")
}

// patterns for PrivateDisplayService
var (
	pattern_goclay_PrivateDisplayService_List_0 = "/v1/displays"

	pattern_goclay_PrivateDisplayService_List_0_builder = func(in *ListRequest) string {
		values := url.Values{}

		u := url.URL{
			Path:     fmt.Sprintf("/v1/displays"),
			RawQuery: values.Encode(),
		}
		return u.String()
	}

	unmarshaler_goclay_PrivateDisplayService_List_0_boundParams = &utilities.DoubleArray{Encoding: map[string]int{}, Base: []int(nil), Check: []int(nil)}

	pattern_goclay_PrivateDisplayService_SetEnabled_0 = "/v1/displays/set-enable"

	pattern_goclay_PrivateDisplayService_SetEnabled_0_builder = func(in *SetEnabledRequest) string {
		values := url.Values{}

		u := url.URL{
			Path:     fmt.Sprintf("/v1/displays/set-enable"),
			RawQuery: values.Encode(),
		}
		return u.String()
	}

	unmarshaler_goclay_PrivateDisplayService_SetEnabled_0_boundParams = &utilities.DoubleArray{Encoding: map[string]int{"": 0}, Base: []int{1, 1, 0}, Check: []int{0, 1, 2}}
)

// marshalers for PrivateDisplayService
var (
	unmarshaler_goclay_PrivateDisplayService_List_0 = func(r *http.Request) func(interface{}) error {
		return func(rif interface{}) error {
			req := rif.(*ListRequest)

			if err := errors.Wrap(runtime.PopulateQueryParameters(req, r.URL.Query(), unmarshaler_goclay_PrivateDisplayService_List_0_boundParams), "couldn't populate query parameters"); err != nil {
				return httpruntime.TransformUnmarshalerError(err)
			}

			return nil
		}
	}

	unmarshaler_goclay_PrivateDisplayService_SetEnabled_0 = func(r *http.Request) func(interface{}) error {
		return func(rif interface{}) error {
			req := rif.(*SetEnabledRequest)

			if err := errors.Wrap(runtime.PopulateQueryParameters(req, r.URL.Query(), unmarshaler_goclay_PrivateDisplayService_SetEnabled_0_boundParams), "couldn't populate query parameters"); err != nil {
				return httpruntime.TransformUnmarshalerError(err)
			}

			inbound, _ := httpruntime.MarshalerForRequest(r)
			if err := errors.Wrap(inbound.Unmarshal(r.Body, &req), "couldn't read request JSON"); err != nil {
				return httptransport.NewMarshalerError(httpruntime.TransformUnmarshalerError(err))
			}
			return nil
		}
	}
)

var _swaggerDef_agent_private_display_v1_service_proto = []byte(`{
  "swagger": "2.0",
  "info": {
    "title": "Display",
    "description": "Display control",
    "version": "1.0"
  },
  "tags": [
    {
      "name": "PrivateDisplayService",
      "description": "Display"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/displays": {
      "get": {
        "description": "list displays",
        "operationId": "PrivateDisplayService_List",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListResponse"
            }
          }
        },
        "tags": [
          "Display"
        ]
      }
    },
    "/v1/displays/set-enable": {
      "post": {
        "description": "Switch display enable",
        "operationId": "PrivateDisplayService_SetEnabled",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SetEnabledResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1SetEnabledRequest"
            }
          }
        ],
        "tags": [
          "Display"
        ]
      }
    }
  },
  "definitions": {
    "v1Display": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "enable": {
          "type": "boolean"
        }
      }
    },
    "v1ListResponse": {
      "type": "object",
      "properties": {
        "displays": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Display"
          }
        }
      }
    },
    "v1SetEnabledRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "enable": {
          "type": "boolean"
        }
      }
    },
    "v1SetEnabledResponse": {
      "type": "object"
    }
  }
}

`)